Don't crash in IE because of missing localStorage for file:// URLs.
IE doesn't seem to support localStorage for file:// URLs. This fix prevents the JS from crashing in IE. Unfortunately this also means that the sidebar width is no longer restored upon navigation when using IE.
The fix checks for the presence of the localStorage, so once IE supports it, already existing HTML files will start to work there as well.