VPRT-818: Fix switching tabs in viewport
The code for smooth scrolling to anchors registers an event handler on
all links to anchors that prevents bubbling by returning false. AUI tabs
use such links, but their click handler is listening on an ancestor
(.tabs-menu), so they rely on the bubbling.

This commit disables the smooth scrolling event handler for AUI tab
links.